+++ js/script.js
@@ -55,9 +55,8 @@ taskForm.addEventListener("submit", function (event) {
 function displayListItems(task) {
   let taskItem = document.createElement("div");
   taskItem.className = "to-do-item";
-  taskItem.innerHTML = 
-        ` <div class="tick-container">
-            <img src="images/check-circle.svg" class="ticks"/>
+  taskItem.innerHTML = ` <div class="tick-container">
+            <img src="images/check-circle.svg" class="iconTicks"/>
           </div>
           <div class="to-do-text-container">
             <p class="task-name">${task.name}</p>
@@ -65,12 +64,14 @@ function displayListItems(task) {
           <div class="importance-container">
             <p class="task-importance">${task.importance}</p>
           </div>
+          <div class="delete-container">
+            <img src="images/delete.svg" class="iconDelete"/>
+          </div>
     `;
 
   toDoListContainer.appendChild(taskItem);
-  const tick = taskItem.querySelector(".ticks");
+  const tick = taskItem.querySelector(".iconTicks");
   tick.addEventListener("click", completeTask);
-  
 }
 
 // Task Completion
@@ -81,7 +82,9 @@ function completeTask(event) {
   const taskItem = event.target.closest(".to-do-item");
 
   if (taskItem) {
-    const taskName = taskItem.querySelector(".to-do-text-container p").textContent;
+    const taskName = taskItem.querySelector(
+      ".to-do-text-container p"
+    ).textContent;
     const correspondingTask = taskList.find((task) => task.name === taskName);
 
     if (correspondingTask) {
@@ -95,10 +98,9 @@ function completeTask(event) {
       // Toggle class to complete
       taskItem.classList.add("complete");
     }
-    console.log(taskItem)
+    console.log(taskItem);
   }
 }
-
 
 /* Task Deletion
   Stuff*/